Added testInstrumentationRunner for dev and prod flavors; Moved LoginActivityTest class into prod flavor and other UI tests to dev flavor
