+++ app/src/androidTestDev/java/com/alex/mapnotes/splash/SplashActivityTest.kt
